[react-article-components] svgs renamed
back-top-topic.svg -> back-to-topic.svg
article-back-to-topic-mobile.svg -> back-to-topic-mobile.svg
article-back-to-top-mobile.svg -> back-to-top-mobile.svg
